# specs/001-pipeline-intelligence/contracts/api.yaml
openapi: 3.1.0
info:
  title: HubSpot Pipeline Intelligence API
  version: 1.0.0
  description: >
    Backend REST API for the Pipeline Intelligence system.
    Handles recommendations, deals, admin configuration,
    Slack interactions, audit trails, and system health.
servers:
  - url: /api/v1
    description: API v1 prefix
security:
  - oidcAuth: []
components:
  securitySchemes:
    oidcAuth:
      type: openIdConnect
      openIdConnectUrl: https://accounts.google.com/.well-known/openid-configuration
  schemas:
...
    Deal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        hubspot_deal_id:
          type: string
        deal_name:
          type: string
        current_stage:
          type: string
        current_stage_name:
          type: string
        pipeline_name:
          type: string
        owner:
          type: object
          properties:
            id:
              type: string
              format: uuid
            display_name:
              type: string
            email:
              type: string
        amount:
          type: number
          nullable: true
        close_date:
          type: string
          format: date
          nullable: true
        thread_count:
          type: integer
        recommendation_count:
          type: integer
        last_synced_at:
          type: string
          format: date-time
...
    AnalyticsOverview:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date-time
            end:
              type: string
              format: date-time
        recommendations:
          type: object
          properties:
            total:
              type: integer
            approved:
              type: integer
            rejected:
              type: integer
            expired:
              type: integer
            pending:
              type: integer
            approval_rate:
              type: number
        rejections_by_reason:
          type: object
          additionalProperties:
            type: integer
        confidence:
          type: object
          properties:
            average:
              type: number
            p50:
              type: number
            p95:
              type: number
        latency:
          type: object
          properties:
            email_to_notification_p50_seconds:
              type: number
            email_to_notification_p95_seconds:
              type: number
            approval_to_crm_p50_seconds:
              type: number
            approval_to_crm_p95_seconds:
              type: number
        deal_coverage:
          type: object
          properties:
            total_active_deals:
              type: integer
            deals_with_threads:
              type: integer
            coverage_percent:
              type: number
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
paths:
...
  /deals:
    get:
      operationId: listDeals
      summary: List monitored deals
      tags: [Deals]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: pipeline_id
          in: query
          schema:
            type: string
            format: uuid
        - name: owner_id
          in: query
          schema:
            type: string
            format: uuid
        - name: has_threads
          in: query
          schema:
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deal'
                  total:
                    type: integer
  /deals/{id}:
    get:
      operationId: getDeal
      summary: Get deal details with threads and recommendations
      tags: [Deals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deal'

  # ── Analytics ────────────────────────────────────────────────
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      summary: Get analytics overview metrics
      description: >
        Returns aggregate metrics: approval rates, rejection reasons,
        confidence distributions, latency percentiles, deal coverage.
        RevOps and admin only.
      tags: [Analytics]
      parameters:
        - name: period_days
          in: query
          schema:
            type: integer
            default: 7
            enum: [7, 14, 30, 90]
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsOverview'
  /analytics/classifications/low-confidence:
    get:
      operationId: listLowConfidenceClassifications
      summary: List below-threshold classifications for RevOps review
      tags: [Analytics]
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: min_confidence
          in: query
          schema:
            type: number
            default: 0.30
        - name: max_confidence
          in: query
          schema:
            type: number
            default: 0.70
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        email_subject:
                          type: string
                        email_excerpt:
                          type: string
                        intent:
                          type: string
                        confidence_score:
                          type: number
                        reasoning:
                          type: string
                        deal_name:
                          type: string
                          nullable: true
                        created_at:
                          type: string
                          format: date-time
                  total:
                    type: integer

  # ── Audit ────────────────────────────────────────────────────
...